senticle50/brexit/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
astroid==1.6.0
billiard==3.3.0.23
boto==2.48.0
boto3==1.5.18
botocore==1.8.32
bz2file==0.98
category-encoders==1.2.6
celery==3.1.25
certifi==2018.1.18
chardet==3.0.4
coverage==4.4.2
cssselect==1.0.1
Cython==0.27.3
decorator==4.1.2
Django==2.0
django-bulk-update==2.2.0
django-endless-pagination==2.0
django-jenkins==0.19.0
django-urltags==0.1
djangorestframework==3.7.3
docutils==0.14
gensim==3.2.0
h5py==2.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.11.1
jmespath==0.9.3
Keras==2.1.3
kombu==3.0.37
lazy-object-proxy==1.3.1
line-profiler==2.1.2
lxml==4.1.1
mccabe==0.6.1
memory-profiler==0.50.0
mockito==1.0.12
nltk==3.2.5
numpy==1.14.0
oauthlib==2.0.6
pandas==0.22.0
parso==0.1.1
patsy==0.5.0
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
PorterStemmer==0.5
prompt-toolkit==1.0.15
protobuf==3.1.0
proxy==0.0.1
psutil==5.4.2
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
pylint==1.8.1
pyquery==1.3.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.5.6
statsmodels==0.8.0
Theano==1.0.1
tqdm==4.19.5
traitlets==4.3.2
tweepy==3.5.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
gunicorn==19.7.1
wrapt==1.10.11
